+++ springy/Template/SmartyDriver.php
@@ -7,16 +7,14 @@
  *
  * @see       http://www.smarty.net/
  *
- * @copyright 2014-2018 Fernando Val
+ * @copyright 2014 Fernando Val
  * @author    Fernando Val <fernando.val@gmail.com>
  * @license   https://github.com/fernandoval/Springy/blob/master/LICENSE MIT
- *
- * @version   1.7.23
  */
 
 namespace Springy\Template;
 
-use Smarty;
+use Smarty\Smarty;
 use Springy\Kernel;
 use Springy\URI;
 
@@ -62,14 +60,18 @@ class SmartyDriver implements TemplateDriverInterface
         $this->tplObj->debugging_ctrl = config_get('template.debugging_ctrl');
         $this->tplObj->use_sub_dirs = config_get('template.use_sub_dirs');
 
-        $this->setCacheDir(config_get('template.template_cached_path'));
-        $this->setTemplateDir([
-            'main' => config_get('template.template_path'),
-            'default' => config_get('template.default_template_path'),
+        $this->tplObj->setCacheDir(config_get('template.template_cached_path'));
+        $this->tplObj->setTemplateDir([
+            config_get('template.template_path'),
+            config_get('template.default_template_path'),
         ]);
-        $this->setCompileDir(config_get('template.compiled_template_path'));
-        $this->setConfigDir(config_get('template.template_config_path'));
-        $this->tplObj->addPluginsDir(config_get('template.template_plugins_path'));
+        $this->tplObj->setCompileDir(config_get('template.compiled_template_path'));
+        $this->tplObj->setConfigDir(config_get('template.template_config_path'));
+
+        foreach (config_get('template.smarty_extensions', []) as $extensionClass) {
+            $this->tplObj->addExtension(new $extensionClass());
+        }
+
         $this->setTemplate($tpl);
     }
 
@@ -216,12 +218,12 @@ class SmartyDriver implements TemplateDriverInterface
 
         // Inicializa as funções personalizadas padrão
         foreach (Kernel::getTemplateFunctions() as $func) {
-            $this->tplObj->registerPlugin($func[0], $func[1], $func[2], $func[3], $func[4]);
+            $this->tplObj->registerPlugin($func[0], $func[1], $func[2], $func[3]);
         }
 
         // Inicializa as funções personalizadas do template
         foreach ($this->templateFuncs as $func) {
-            $this->tplObj->registerPlugin($func[0], $func[1], $func[2], $func[3], $func[4]);
+            $this->tplObj->registerPlugin($func[0], $func[1], $func[2], $func[3]);
         }
 
         return $this->tplObj->fetch(
@@ -243,18 +245,26 @@ class SmartyDriver implements TemplateDriverInterface
         // Se o nome do template não foi informado, define como relativo à controladora atual
         if ($tpl === null) {
             // Pega o caminho relativo da página atual
-            $path = URI::relativePathPage(true);
-            $this->setTemplate($path . (empty($path) ? '' : DIRECTORY_SEPARATOR) . URI::getControllerClass());
+            $path = implode(
+                DS,
+                array_filter(
+                    array_merge(
+                        Kernel::getTemplatePrefix(),
+                        [URI::relativePathPage()]
+                    )
+                )
+            );
+            $this->setTemplate($path . (empty($path) ? '' : DS) . URI::getControllerClass());
 
             return;
         }
 
-        $this->templateName = ((is_array($tpl)) ? implode(DIRECTORY_SEPARATOR, $tpl) : $tpl);
+        $this->templateName = ((is_array($tpl)) ? implode(DS, $tpl) : $tpl);
 
         $compile = '';
         if (!is_null($tpl)) {
-            $compile = is_array($tpl) ? implode(DIRECTORY_SEPARATOR, $tpl) : $tpl;
-            $compile = substr($compile, 0, strrpos(DIRECTORY_SEPARATOR, $compile));
+            $compile = is_array($tpl) ? implode(DS, $tpl) : $tpl;
+            $compile = substr($compile, 0, strrpos(DS, $compile));
         }
 
         $this->setCompileDir(config_get('template.compiled_template_path') . $compile);
@@ -311,13 +321,12 @@ class SmartyDriver implements TemplateDriverInterface
      * @param string       $name        defines the name of the plugin.
      * @param string|array $callback    defines the callback.
      * @param mixed        $cacheable
-     * @param mixed        $cache_attrs
      *
      * @return void
      */
-    public function registerPlugin($type, $name, $callback, $cacheable = null, $cache_attrs = null)
+    public function registerPlugin($type, $name, $callback, $cacheable = null)
     {
-        $this->templateFuncs[] = [$type, $name, $callback, $cacheable, $cache_attrs];
+        $this->templateFuncs[] = [$type, $name, $callback, $cacheable];
     }
 
     /**
@@ -411,8 +420,8 @@ class SmartyDriver implements TemplateDriverInterface
             return '#';
         }
 
-        $srcPath = config_get('system.assets_source_path') . DIRECTORY_SEPARATOR . $params['file'];
-        $filePath = config_get('system.assets_path') . DIRECTORY_SEPARATOR . $params['file'];
+        $srcPath = config_get('system.assets_source_path') . DS . $params['file'];
+        $filePath = config_get('system.assets_path') . DS . $params['file'];
         $fileURI = config_get('uri.assets_dir') . '/' . $params['file'];
         $get = [];