+++ springy/Template/TemplateDriverInterface.php
@@ -140,11 +140,10 @@ interface TemplateDriverInterface
      * @param string       $name        defines the name of the plugin.
      * @param string|array $callback    defines the callback.
      * @param mixed        $cacheable
-     * @param mixed        $cache_attrs
      *
      * @return void
      */
-    public function registerPlugin($type, $name, $callback, $cacheable = null, $cache_attrs = null);
+    public function registerPlugin($type, $name, $callback, $cacheable = null);
 
     /**
      * Clears the value of an assigned variable.